fixed bugs in steady state calculation and equilibrium conditions
Bug 1: Labor can not be normalized to one (as in SW) since it violates
FOC for labor in steady state.
Proof: This was mentioned in appendix (page 58) but not implemented in
the code.

Bug 2: missing term z_{t-1} and incorrect coefficient for term z_t in
Phillips curve for wages
Proof: If we take wage Phillips curve (equation 13 in SW) and replace
pi_t with pi_t+z_t then we don't get the same equation as 8.106.